# Conflicts:
#	src/Common/MasterDataProxies/src/BaseProxy.cs
#	src/Common/MasterDataProxies/src/Interfaces/IRaptorProxy.cs
#	src/Common/MasterDataProxies/src/RaptorProxy.cs
#	src/Common/Productivity3DModels/src/Models/FilterResult.cs
#	src/service/MockProjectWebApi/Controllers/MockProductivity3dV2CompactionController.cs
#	src/service/Scheduler/VSS.Productivity3D.Scheduler.sln

+++ src/service/Project/src/ProjectWebApiCommon/Helpers/DataOceanHelper.cs
@@ -5,12 +5,16 @@ using System.Net;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
+using VSS.Common.Abstractions.Configuration;
 using VSS.Common.Abstractions.Extensions;
 using VSS.Common.Abstractions.Http;
+using VSS.Common.Exceptions;
 using VSS.DataOcean.Client;
 using VSS.MasterData.Models.Handlers;
+using VSS.MasterData.Models.ResultHandling.Abstractions;
 using VSS.MasterData.Project.WebAPI.Common.Models;
 using VSS.MasterData.Project.WebAPI.Common.Utilities;
+using VSS.Productivity3D.Project.Abstractions.Models.ResultsHandling;
 using VSS.WebApi.Common;
 
 namespace VSS.MasterData.Project.WebAPI.Common.Helpers
@@ -30,44 +34,55 @@ namespace VSS.MasterData.Project.WebAPI.Common.Helpers
     /// </summary>
     public static async Task WriteFileToDataOcean(
       Stream fileContents, string rootFolder, string customerUid, string projectUid,
-      string pathAndFileName, bool isSurveyedSurface, DateTime? surveyedUtc,
+      string dataOceanFileName, bool isSurveyedSurface, DateTime? surveyedUtc,
       ILogger log, IServiceExceptionHandler serviceExceptionHandler, IDataOceanClient dataOceanClient,
-      ITPaaSApplicationAuthentication authn)
+      ITPaaSApplicationAuthentication authn, Guid fileUid, IConfigurationStore configStore)
     {
-      var customHeaders = authn.CustomHeaders();
-      var dataOceanPath = DataOceanPath(rootFolder, customerUid, projectUid);
-      string dataOceanFileName = Path.GetFileName(pathAndFileName);
-
-      //TODO: DataOcean has versions of files. We should leverage that rather than appending surveyed UTC to file name.
-      if (isSurveyedSurface && surveyedUtc != null) // validation should prevent this
-        dataOceanFileName = dataOceanFileName.IncludeSurveyedUtcInName(surveyedUtc.Value);
-
-      bool ccPutFileResult = false;
-      bool folderAlreadyExists = false;
-      try
+      var dataOceanEnabled = configStore.GetValueBool("ENABLE_DATA_OCEAN", false);
+      if (dataOceanEnabled)
       {
+        //Check file name starts with a Guid
+        if (!dataOceanFileName.StartsWith(fileUid.ToString()))
+        {
+          throw new ServiceException(HttpStatusCode.InternalServerError,
+            new ContractExecutionResult(ContractExecutionStatesEnum.InternalProcessingError,
+              $"Invalid DataOcean file name {dataOceanFileName}"));
+        }
+
+        var customHeaders = authn.CustomHeaders();
+        var dataOceanPath = DataOceanPath(rootFolder, customerUid, projectUid);
+
+        bool ccPutFileResult = false;
+        bool folderAlreadyExists = false;
+        try
+        {
+          log.LogInformation(
+            $"WriteFileToDataOcean: dataOceanPath {dataOceanPath} dataOceanFileName {dataOceanFileName}");
+          folderAlreadyExists = await dataOceanClient.FolderExists(dataOceanPath, customHeaders);
+          if (folderAlreadyExists == false)
+            await dataOceanClient.MakeFolder(dataOceanPath, customHeaders);
+
+          // this does an upsert
+          ccPutFileResult = await dataOceanClient.PutFile(dataOceanPath, dataOceanFileName, fileContents, customHeaders);
+        }
+        catch (Exception e)
+        {
+          serviceExceptionHandler.ThrowServiceException(HttpStatusCode.InternalServerError, 57, "dataOceanClient.PutFile",
+            e.Message);
+        }
+
+        if (ccPutFileResult == false)
+        {
+          serviceExceptionHandler.ThrowServiceException(HttpStatusCode.InternalServerError, 116);
+        }
+
         log.LogInformation(
-          $"WriteFileToDataOcean: dataOceanPath {dataOceanPath} dataOceanFileName {dataOceanFileName}");
-        folderAlreadyExists = await dataOceanClient.FolderExists(dataOceanPath, customHeaders);
-        if (folderAlreadyExists == false)
-          await dataOceanClient.MakeFolder(dataOceanPath, customHeaders);
-
-        // this does an upsert
-        ccPutFileResult = await dataOceanClient.PutFile(dataOceanPath, dataOceanFileName, fileContents, customHeaders);
+          $"WriteFileToDataOcean: dataOceanFileName {dataOceanFileName} written to DataOcean. folderAlreadyExists {folderAlreadyExists}");
       }
-      catch (Exception e)
+      else
       {
-        serviceExceptionHandler.ThrowServiceException(HttpStatusCode.InternalServerError, 57, "dataOceanClient.PutFile",
-          e.Message);
+        log.LogInformation($"WriteFileToDataOcean: File not saved. DataOcean disabled");
       }
-
-      if (ccPutFileResult == false)
-      {
-        serviceExceptionHandler.ThrowServiceException(HttpStatusCode.InternalServerError, 116);
-      }
-
-      log.LogInformation(
-        $"WriteFileToDataOcean: dataOceanFileName {dataOceanFileName} written to DataOcean. folderAlreadyExists {folderAlreadyExists}");
     }
 
     /// <summary>
@@ -77,32 +92,40 @@ namespace VSS.MasterData.Project.WebAPI.Common.Helpers
     public static async Task<ImportedFileInternalResult> DeleteFileFromDataOcean(
       string fileName, string rootFolder, string customerUid, Guid projectUid, Guid importedFileUid, 
       ILogger log, IServiceExceptionHandler serviceExceptionHandler, 
-      IDataOceanClient dataOceanClient, ITPaaSApplicationAuthentication authn)
+      IDataOceanClient dataOceanClient, ITPaaSApplicationAuthentication authn, IConfigurationStore configStore)
     {
-      var dataOceanPath = DataOceanPath(rootFolder, customerUid, projectUid.ToString());
-      var fullFileName = $"{dataOceanPath}{Path.DirectorySeparatorChar}{fileName}";
-      log.LogInformation($"DeleteFileFromDataOcean: fullFileName {JsonConvert.SerializeObject(fullFileName)}");
+      var dataOceanEnabled = configStore.GetValueBool("ENABLE_DATA_OCEAN", false);
+      if (dataOceanEnabled)
+      {
+        var dataOceanPath = DataOceanPath(rootFolder, customerUid, projectUid.ToString());
+        var fullFileName = $"{dataOceanPath}{Path.DirectorySeparatorChar}{fileName}";
+        log.LogInformation($"DeleteFileFromDataOcean: fullFileName {JsonConvert.SerializeObject(fullFileName)}");
 
-      var customHeaders = authn.CustomHeaders();
-      bool ccDeleteFileResult = false;
-      try
-      {
-        ccDeleteFileResult = await dataOceanClient.DeleteFile(fullFileName, customHeaders);
+        var customHeaders = authn.CustomHeaders();
+        bool ccDeleteFileResult = false;
+        try
+        {
+          ccDeleteFileResult = await dataOceanClient.DeleteFile(fullFileName, customHeaders);
+        }
+        catch (Exception e)
+        {
+          log.LogError(e, $"DeleteFileFromDataOcean failed for {fileName} (importedFileUid:{importedFileUid}) with exception {e.Message}");
+          return ImportedFileInternalResult.CreateImportedFileInternalResult(HttpStatusCode.InternalServerError, 57, "dataOceanClient.DeleteFile", e.Message);
+        }
+
+        if (ccDeleteFileResult == false)
+        {
+          log.LogWarning(
+            $"DeleteFileFromDataOcean failed to delete {fileName} (importedFileUid:{importedFileUid}).");
+          //Not an error if it doesn't delete the file?
+          //return ImportedFileInternalResult.CreateImportedFileInternalResult(HttpStatusCode.InternalServerError, 117);
+        }
       }
-      catch (Exception e)
+      else
       {
-        log.LogError(e, $"DeleteFileFromDataOcean failed for {fileName} (importedFileUid:{importedFileUid}) with exception {e.Message}");
-        return ImportedFileInternalResult.CreateImportedFileInternalResult(HttpStatusCode.InternalServerError, 57, "dataOceanClient.DeleteFile", e.Message);
+        log.LogInformation($"DeleteFileFromDataOcean: File not deleted. DataOcean disabled");
       }
 
-      if (ccDeleteFileResult == false)
-      {
-        log.LogWarning(
-          $"DeleteFileFromDataOcean failed to delete {fileName} (importedFileUid:{importedFileUid}).");
-        //Not an error if it doesn't delete the file?
-        //return ImportedFileInternalResult.CreateImportedFileInternalResult(HttpStatusCode.InternalServerError, 117);
-      }
-    
       return null;
     }
